backend/requirements.txt
Django
django-ninja
django-storages[s3]
boto3-stubs[s3]
django-stubs-ext
django-types
gunicorn
psycopg2-binary
pydantic
pydantic-settings
whitenoise-spa
Pillow
